Added onClick function to Enroll button
The Enroll button in the SingleCoursesTwo component did not have any functionality. In this commit, an onClick handler was added to the Enroll button so that it triggers the 'request access' function when clicked. This change will allow users to enroll for free courses when they click on the Enroll button.